Added autoencoder to ann module
